requirements.txt
# ============================================================
# RAG Q&A System - Dependencies
# ============================================================

# Core LangChain ecosystem
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-huggingface==0.0.3

# OpenAI
openai==1.40.0

# Vector Store
faiss-cpu==1.10.0

# PDF Processing
pypdf==4.3.1
pymupdf==1.24.9

# API Backend
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

# Frontend
streamlit==1.37.1

# Environment & Config
python-dotenv==1.0.1

# HuggingFace (fallback LLM & embeddings)
transformers==4.44.0
sentence-transformers==3.0.1
torch==2.4.0

# Utilities
tiktoken==0.7.0
httpx==0.27.0
aiofiles==24.1.0

# Logging
loguru==0.7.2

# Dev tools (optional)
pytest==8.3.2
black==24.8.0
